Make R type `integer' correspond to C type `int' in .C calls.
git-svn-id: https://svn.r-project.org/R/trunk@4032 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/appl/rowsum.c
@@ -15,7 +15,7 @@
 */
 
 void rowsum(dim, na_x, x, group)
-long    *dim;
+int     *dim;
 double  *na_x;
 double  *x,
 	*group;
